feat: :art: creation of the navigation bar and code optimization
Refactoring and improvement of the navigation bar  In this revision, I made a number of changes to the NavBar component to improve its readability and performance. Replaced inline style classes with tailwindcss for better code organization and maintenance. Additionally, the Container component was applied to wrap the navigation bar and semantic tags were implemented for better accessibility. Fixed typos and adjusted styles for a more consistent look. In addition, search, cart counter and user menu functions were incorporated. Optimized the use of the 'Redressed' font to reduce page load. This version provides an improved user experience and cleaner code.