feat: :sparkles: Pass product object directly to ProductDetails component
Modified the Product component to pass the product object directly to the ProductDetails component instead of fetching the product based on productId. This is done to avoid making an additional data request as the product details are already defined in the utility file. Ensured that the product object is properly passed to the ProductDetails component.